feat: change status and download

+++ api/internal/logic/file/deletefilelogic.go
@@ -2,16 +2,17 @@ package file
 
 import (
 	"context"
-	"github.com/suyuan32/simple-admin-core/api/common/errorx"
-	"github.com/suyuan32/simple-admin-file/api/internal/model"
-	"github.com/zeromicro/go-zero/rest/httpx"
+	"encoding/json"
 	"net/http"
 	"os"
 
+	"github.com/suyuan32/simple-admin-file/api/internal/model"
 	"github.com/suyuan32/simple-admin-file/api/internal/svc"
 	"github.com/suyuan32/simple-admin-file/api/internal/types"
 
+	"github.com/zeromicro/go-zero/core/errorx"
 	"github.com/zeromicro/go-zero/core/logx"
+	"github.com/zeromicro/go-zero/rest/httpx"
 )
 
 type DeleteFileLogic struct {
@@ -34,14 +35,29 @@ func (l *DeleteFileLogic) DeleteFile(req *types.IdReq) (resp *types.SimpleMsg, e
 	if result.Error != nil {
 		return nil, httpx.NewApiError(http.StatusInternalServerError, errorx.DatabaseError)
 	}
+	if result.RowsAffected == 0 {
+		return nil, httpx.NewApiErrorWithoutMsg(http.StatusNotFound)
+	}
+
+	// judge the user is admin or owner
+	// 只有管理员和拥有者能删除信息
+	if l.ctx.Value("roleId").(json.Number).String() != "1" && l.ctx.Value("userId").(string) != target.UserUUID {
+		return nil, httpx.NewApiErrorWithoutMsg(http.StatusUnauthorized)
+	}
+
 	if target.Status {
 		err = os.Remove(l.svcCtx.Config.UploadConf.PublicStorePath + target.Path)
 		if err != nil {
 			return nil, httpx.NewApiErrorWithoutMsg(http.StatusInternalServerError)
 		}
+	} else {
+		err = os.Remove(l.svcCtx.Config.UploadConf.PrivateStorePath + target.Path)
+		if err != nil {
+			return nil, httpx.NewApiErrorWithoutMsg(http.StatusInternalServerError)
+		}
 	}
 
-	result = l.svcCtx.DB.Delete(target)
+	result = l.svcCtx.DB.Delete(&target)
 	if result.Error != nil {
 		return nil, httpx.NewApiError(http.StatusInternalServerError, errorx.DatabaseError)
 	}
